+++ src/ambition/7kaaInterface/draw.cc
@@ -26,6 +26,7 @@
 #define _AMBITION_IMPLEMENTATION
 #include "draw.hh"
 
+#include "OCONFIG.h"
 #include "OFIRM.h"
 #include "ONATIONA.h"
 #include "OTOWN.h"
@@ -36,6 +37,7 @@
 #include "Ambition_minimap.hh"
 #include "Ambition_remote.hh"
 #include "Ambition_unit.hh"
+#include "Ambition_user_interface.hh"
 #include "Ambition_vga.hh"
 
 
@@ -384,6 +386,24 @@ void modeInformation(
   Ambition::drawModeInformation();
 }
 
+void outsideLeadershipIcon(
+  Unit* _7kaaUnit
+) {
+  if (!Ambition::config.enhancementsAvailable()) {
+    return;
+  }
+
+  if (!_7kaaUnit->is_own() && !config.show_ai_info) {
+    return;
+  }
+
+  if (Ambition::Unit::canReceiveLeadershipBonus(_7kaaUnit)
+      && !Ambition::Unit::isReceivingLeadershipBonus(_7kaaUnit)
+  ) {
+    Ambition::drawOutsideLeadershipIcon(_7kaaUnit);
+  }
+}
+
 void printGameSpeed(
   const int speed
 ) {
@@ -392,6 +412,27 @@ void printGameSpeed(
   }
 
   Ambition::displayGameSpeed(speed);
+}
+
+bool printLeadershipStatus(
+  Unit* _7kaaUnit,
+  const int top,
+  const int refreshFlag
+) {
+  if (!Ambition::config.enhancementsAvailable()) {
+    return false;
+  }
+
+  if (!_7kaaUnit->is_own() && !config.show_ai_info) {
+    return false;
+  }
+
+  if (!Ambition::Unit::canReceiveLeadershipBonus(_7kaaUnit)) {
+    return false;
+  }
+
+  Ambition::printLeadershipStatus(_7kaaUnit, top, refreshFlag);
+  return true;
 }
 
 void printSyncError(